requirements.txt
streamlit==1.40.0
pillow==10.0.0
reportlab==4.0.9
python-dotenv==1.0.0
streamlit-super-slider
captcha==0.5.0
